+++ .devcontainer/post-install.sh
@@ -1,3 +1,4 @@
 #! /bin/zsh
 
 poetry install --all-extras --with dev,doc
+julia -e "using Pkg; Pkg.add(\"PyCall\"); Pkg.add(\"MultiFloats\")"