+++ mrmustard/math/lattice/strategies/compactFock/inputValidation.py
@@ -1,17 +1,21 @@
 """
 This module contains helper functions that are used in
-compactFock_diagonal_amps.py, compactFock_diagonal_grad.py, compactFock_1leftoverMode_amps.py and compactFock_1leftoverMode_grad.py
+diagonal_amps.py, diagonal_grad.py, singleLeftoverMode_amps.py and singleLeftoverMode_grad.py
 to validate the input provided by the user.
 """
 
 from typing import Iterable
 import numpy as np
-from mrmustard.math.numba.compactFock_diagonal_amps import fock_representation_diagonal_amps
-from mrmustard.math.numba.compactFock_diagonal_grad import fock_representation_diagonal_grad
-from mrmustard.math.numba.compactFock_1leftoverMode_amps import (
+from mrmustard.math.lattice.strategies.compactFock.diagonal_amps import (
+    fock_representation_diagonal_amps,
+)
+from mrmustard.math.lattice.strategies.compactFock.diagonal_grad import (
+    fock_representation_diagonal_grad,
+)
+from mrmustard.math.lattice.strategies.compactFock.singleLeftoverMode_amps import (
     fock_representation_1leftoverMode_amps,
 )
-from mrmustard.math.numba.compactFock_1leftoverMode_grad import (
+from mrmustard.math.lattice.strategies.compactFock.singleLeftoverMode_grad import (
     fock_representation_1leftoverMode_grad,
 )
 from thewalrus._hafnian import input_validation
@@ -19,7 +23,7 @@ from thewalrus._hafnian import input_validation
 
 def hermite_multidimensional_diagonal(A, B, G0, cutoffs, rtol=1e-05, atol=1e-08):
     """
-    Validation of user input for mrmustard.math.tensorflow.hermite_renormalized_diagonal
+    Validation of user input for mrmustard.math.tensorflow_wrapper.hermite_renormalized_diagonal
     """
     input_validation(A, atol=atol, rtol=rtol)
     if A.shape[0] != B.shape[0]:
@@ -36,7 +40,7 @@ def hermite_multidimensional_diagonal(A, B, G0, cutoffs, rtol=1e-05, atol=1e-08)
 
 def grad_hermite_multidimensional_diagonal(A, B, G0, arr0, arr2, arr1010, arr1001, arr1):
     """
-    Validation of user input for gradients of mrmustard.math.tensorflow.hermite_renormalized_diagonal
+    Validation of user input for gradients of mrmustard.math.tensorflow_wrapper.hermite_renormalized_diagonal
     """
     if A.shape[0] != B.shape[0]:
         raise ValueError("The matrix A and vector B have incompatible dimensions")
@@ -50,7 +54,7 @@ def grad_hermite_multidimensional_diagonal(A, B, G0, arr0, arr2, arr1010, arr100
 
 def hermite_multidimensional_1leftoverMode(A, B, G0, cutoffs, rtol=1e-05, atol=1e-08):
     """
-    Validation of user input for mrmustard.math.tensorflow.hermite_renormalized_1leftoverMode
+    Validation of user input for mrmustard.math.tensorflow_wrapper.hermite_renormalized_1leftoverMode
     """
     input_validation(A, atol=atol, rtol=rtol)
     if A.shape[0] != B.shape[0]:
@@ -69,7 +73,7 @@ def hermite_multidimensional_1leftoverMode(A, B, G0, cutoffs, rtol=1e-05, atol=1
 
 def grad_hermite_multidimensional_1leftoverMode(A, B, G0, arr0, arr2, arr1010, arr1001, arr1):
     """
-    Validation of user input for gradients of mrmustard.math.tensorflow.hermite_renormalized_1leftoverMode
+    Validation of user input for gradients of mrmustard.math.tensorflow_wrapper.hermite_renormalized_1leftoverMode
     """
     if A.shape[0] != B.shape[0]:
         raise ValueError("The matrix A and vector B have incompatible dimensions")