+++ mrmustard/math/tensorflow_wrapper.py
@@ -23,14 +23,13 @@ import tensorflow_probability as tfp
 from mrmustard import settings
 from mrmustard.math.autocast import Autocast
 from mrmustard.math.lattice import strategies
-from mrmustard.math.numba.compactFock_inputValidation import (
+from mrmustard.math.lattice.strategies.compactFock.inputValidation import (
     grad_hermite_multidimensional_1leftoverMode,
     grad_hermite_multidimensional_diagonal,
     hermite_multidimensional_1leftoverMode,
     hermite_multidimensional_diagonal,
 )
-from mrmustard.typing import Tensor, Trainable
-
+from mrmustard.utils.typing import Tensor, Trainable
 from .math_interface import MathInterface
 
 
@@ -81,7 +80,10 @@ class TFMath(MathInterface):
         return tf.clip_by_value(array, a_min, a_max)
 
     def concat(self, values: Sequence[tf.Tensor], axis: int) -> tf.Tensor:
-        return tf.concat(values, axis)
+        if any(tf.rank(v) == 0 for v in values):
+            return tf.stack(values, axis)
+        else:
+            return tf.concat(values, axis)
 
     def conj(self, array: tf.Tensor) -> tf.Tensor:
         return tf.math.conj(array)
@@ -341,9 +343,9 @@ class TFMath(MathInterface):
 
     # TODO: is a wrapper class better?
     @staticmethod
-    def DefaultEuclideanOptimizer() -> tf.keras.optimizers.Optimizer:
+    def DefaultEuclideanOptimizer() -> tf.keras.optimizers.legacy.Optimizer:
         r"""Default optimizer for the Euclidean parameters."""
-        return tf.keras.optimizers.Adam(learning_rate=0.001)
+        return tf.keras.optimizers.legacy.Adam(learning_rate=0.001)
 
     def value_and_gradients(
         self, cost_fn: Callable, parameters: List[Trainable]
@@ -380,8 +382,26 @@ class TFMath(MathInterface):
         Returns:
             The renormalized Hermite polynomial of given shape.
         """
+
+        precision_bits = settings.PRECISION_BITS_HERMITE_POLY
+
         _A, _B, _C = self.asnumpy(A), self.asnumpy(B), self.asnumpy(C)
-        G = strategies.vanilla(tuple(shape), _A, _B, _C)
+
+        if precision_bits == 128:  # numba
+            G = strategies.vanilla(tuple(shape), _A, _B, _C)
+        else:  # julia
+            # The following import must come after running "jl = Julia(compiled_modules=False)" in settings.py
+            from julia import Main as Main_julia  # pylint: disable=import-outside-toplevel
+
+            _A, _B, _C = (
+                _A.astype(np.complex128),
+                _B.astype(np.complex128),
+                _C.astype(np.complex128),
+            )
+
+            G = Main_julia.Vanilla.vanilla(
+                _A, _B, _C.item(), np.array(shape, dtype=np.int64), precision_bits
+            )
 
         def grad(dLdGconj):
             dLdA, dLdB, dLdC = strategies.vanilla_vjp(G, _C, np.conj(dLdGconj))
@@ -433,7 +453,7 @@ class TFMath(MathInterface):
         return G, grad
 
     def reorder_AB_bargmann(self, A: tf.Tensor, B: tf.Tensor) -> Tuple[tf.Tensor, tf.Tensor]:
-        r"""In mrmustard.math.numba.compactFock~ dimensions of the Fock representation are ordered like [mode0,mode0,mode1,mode1,...]
+        r"""In mrmustard.math.compactFock.compactFock~ dimensions of the Fock representation are ordered like [mode0,mode0,mode1,mode1,...]
         while in mrmustard.physics.bargmann the ordering is [mode0,mode1,...,mode0,mode1,...]. Here we reorder A and B.
         """
         ordering = np.arange(2 * A.shape[0] // 2).reshape(2, -1).T.flatten()
@@ -445,7 +465,7 @@ class TFMath(MathInterface):
     def hermite_renormalized_diagonal(
         self, A: tf.Tensor, B: tf.Tensor, C: tf.Tensor, cutoffs: Tuple[int]
     ) -> tf.Tensor:
-        r"""First, reorder A and B parameters of Bargmann representation to match conventions in mrmustard.math.numba.compactFock~
+        r"""First, reorder A and B parameters of Bargmann representation to match conventions in mrmustard.math.compactFock.compactFock~
         Then, calculate the required renormalized multidimensional Hermite polynomial.
         """
         A, B = self.reorder_AB_bargmann(A, B)
@@ -471,16 +491,34 @@ class TFMath(MathInterface):
         Returns:
             The renormalized Hermite polynomial.
         """
-        poly0, poly2, poly1010, poly1001, poly1 = tf.numpy_function(
-            hermite_multidimensional_diagonal, [A, B, C, cutoffs], [A.dtype] * 5
-        )
+        A, B, C = self.asnumpy(A), self.asnumpy(B), self.asnumpy(C)
+        precision_bits = settings.PRECISION_BITS_HERMITE_POLY
+
+        if precision_bits == 128:  # numba (complex128)
+            poly0, poly2, poly1010, poly1001, poly1 = tf.numpy_function(
+                hermite_multidimensional_diagonal, [A, B, C, cutoffs], [A.dtype] * 5
+            )
+        else:  # julia (higher precision than complex128)
+            # The following import must come after running "jl = Julia(compiled_modules=False)" in settings.py
+            from julia import Main as Main_julia  # pylint: disable=import-outside-toplevel
+
+            poly0, poly2, poly1010, poly1001, poly1 = Main_julia.DiagonalAmps.fock_diagonal_amps(
+                A, B, C.item(), tuple(cutoffs), precision_bits
+            )
 
         def grad(dLdpoly):
-            dpoly_dC, dpoly_dA, dpoly_dB = tf.numpy_function(
-                grad_hermite_multidimensional_diagonal,
-                [A, B, C, poly0, poly2, poly1010, poly1001, poly1],
-                [poly0.dtype] * 3,
-            )
+            if precision_bits == 128:  # numba (complex128)
+                dpoly_dC, dpoly_dA, dpoly_dB = tf.numpy_function(
+                    grad_hermite_multidimensional_diagonal,
+                    [A, B, C.item(), poly0, poly2, poly1010, poly1001, poly1],
+                    [poly0.dtype] * 3,
+                )
+            else:  # julia (higher precision than complex128)
+                dpoly_dC = poly0 / C.item()
+                dpoly_dA, dpoly_dB = Main_julia.DiagonalGrad.fock_diagonal_grad(
+                    A, B, poly0, poly2, poly1010, poly1001, poly1, precision_bits
+                )
+
             ax = tuple(range(dLdpoly.ndim))
             dLdA = self.sum(dLdpoly[..., None, None] * self.conj(dpoly_dA), axes=ax)
             dLdB = self.sum(dLdpoly[..., None] * self.conj(dpoly_dB), axes=ax)
@@ -519,16 +557,40 @@ class TFMath(MathInterface):
         Returns:
             The renormalized Hermite polynomial.
         """
-        poly0, poly2, poly1010, poly1001, poly1 = tf.numpy_function(
-            hermite_multidimensional_1leftoverMode, [A, B, C, cutoffs], [A.dtype] * 5
-        )
+        A, B, C = self.asnumpy(A), self.asnumpy(B), self.asnumpy(C)
+        precision_bits = settings.PRECISION_BITS_HERMITE_POLY
+
+        if precision_bits == 128:  # numba (complex128)
+            poly0, poly2, poly1010, poly1001, poly1 = tf.numpy_function(
+                hermite_multidimensional_1leftoverMode, [A, B, C.item(), cutoffs], [A.dtype] * 5
+            )
+        else:  # julia (higher precision than complex128)
+            # The following import must come after running "jl = Julia(compiled_modules=False)" in settings.py
+            from julia import Main as Main_julia  # pylint: disable=import-outside-toplevel
+
+            (
+                poly0,
+                poly2,
+                poly1010,
+                poly1001,
+                poly1,
+            ) = Main_julia.LeftoverModeAmps.fock_1leftoverMode_amps(
+                A, B, C.item(), tuple(cutoffs), precision_bits
+            )
 
         def grad(dLdpoly):
-            dpoly_dC, dpoly_dA, dpoly_dB = tf.numpy_function(
-                grad_hermite_multidimensional_1leftoverMode,
-                [A, B, C, poly0, poly2, poly1010, poly1001, poly1],
-                [poly0.dtype] * 3,
-            )
+            if precision_bits == 128:  # numba (complex128)
+                dpoly_dC, dpoly_dA, dpoly_dB = tf.numpy_function(
+                    grad_hermite_multidimensional_1leftoverMode,
+                    [A, B, C, poly0, poly2, poly1010, poly1001, poly1],
+                    [poly0.dtype] * 3,
+                )
+            else:  # julia (higher precision than complex128)
+                dpoly_dC = poly0 / C.item()
+                dpoly_dA, dpoly_dB = Main_julia.LeftoverModeGrad.fock_1leftoverMode_grad(
+                    A, B, poly0, poly2, poly1010, poly1001, poly1, precision_bits
+                )
+
             ax = tuple(range(dLdpoly.ndim))
             dLdA = self.sum(dLdpoly[..., None, None] * self.conj(dpoly_dA), axes=ax)
             dLdB = self.sum(dLdpoly[..., None] * self.conj(dpoly_dB), axes=ax)